Added a per file backup in an attempt to remove the incremental zip and backup less info and write less to storage each time.
But the file operations ended up being too slow.

I suspect that looking at the file modification times uses DocumentFiles is the slow part. But the file and open and copy operations to DocumentFile also seemed to be slow.

So in the next step I'll just delete the file remove. It's here in history if you want to restore it.

An alternative would be to make it so that there's 1000 audio file directories instead of 100 that would make the amount written lower. But I am not sure it's worth it. I would say don't waste your time.